RSpec tests, code refactors accordingly and Rubocop
